requirements.txt
APScheduler==3.6.3
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
backports.zoneinfo==0.2.1
binance-connector==1.0.0
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
constantly==15.1.0
cryptography==3.4.7
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==20.0.1
python-dotenv==0.19.0
python-telegram-bot==13.7
pytz==2021.1
requests==2.25.1
schedule==1.1.0
service-identity==21.1.0
six==1.16.0
tornado==6.1
Twisted==21.2.0
twisted-iocpsupport==1.0.1
txaio==21.2.1
tzdata==2021.1
tzlocal==3.0
urllib3==1.26.5
zope.interface==5.4.0
